+++ lib/widget/preferred_size.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/widgets.dart';
+
+extension PreferredSizeWidget on Widget {
+  PreferredSize preferredSize(Size size) =>
+      PreferredSize(preferredSize: size, child: this);
+}